Assistant checkpoint: Fix signedness comparison warnings in lexer.cpp
Assistant generated file changes:
- ai_language/src/lexer.cpp: Fix signedness comparison warnings, Fix signedness comparison warning in isAtEnd method

---

User prompt:

-- Configuring done (0.0s)
-- Generating done (0.0s)
[  6%] Building CXX object CMakeFiles/ai_language_lib.dir/src/interpreters/DLInterpreter.cpp.o
[ 12%] Building CXX object CMakeFiles/ai_language_lib.dir/src/lexer.cpp.o
/home/runner/workspace/ai_language/src/lexer.cpp: In member function ‘char ai_language::Lexer::peekNext() const’:
/home/runner/workspace/ai_language/src/lexer.cpp:212:22: warning: comparison of integer expressions of different signedness: ‘int’ and ‘std::__cxx11::basic_string<char>::size_type’ {aka ‘long unsigned int’} [-Wsign-compare]
  212 |     if (_current + 1 >= _source.length()) return '\0';
      |         ~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~
/home/runner/workspace/ai_language/src/lexer.cpp: In member function ‘bool ai_language::Lexer::isAtEnd() const’:
/home/runner/workspace/ai_language/src/lexer.cpp:230:21: warning: comparison of integer expressions of different signedness: ‘const int’ and ‘std::__cxx11::basic_string<char>::size_type’ {aka ‘long unsigned int’} [-Wsign-compare]
  230 |     return _current >= _source.length();
      |            ~~~~~~~~~^~~~~~~~~~~~~~~~~~~
[ 18%] Linking CXX static library libai_language_lib.a
[ 75%] Built target ai_language_lib
[ 87%] Built target show_time
[ 93%] Linking CXX executable /home/runner/workspace/ai_language/ai_lang

+++ ai_language/src/lexer.cpp
@@ -209,7 +209,7 @@ char Lexer::peek() const {
 }
 
 char Lexer::peekNext() const {
-    if (_current + 1 >= _source.length()) return '\0';
+    if (_current + 1 >= static_cast<int>(_source.length())) return '\0';
     return _source[_current + 1];
 }
 
@@ -227,7 +227,7 @@ void Lexer::addToken(TokenType type, const std::string& value) {
 }
 
 bool Lexer::isAtEnd() const {
-    return _current >= _source.length();
+    return _current >= static_cast<int>(_source.length());
 }
 
 void Lexer::error(const std::string& message) {
